新增 redis 服务，注册、登陆、获取验证码接口

+++ GSQ-forum-dao/src/main/java/com/sixtyrobbers/GSQ/fourm/dao/entity/fourm/dbdo/ArticleCommentDO.java
@@ -67,6 +67,6 @@ public class ArticleCommentDO  implements Serializable {
 	 * 有效标志位：0：有效；1：无效
 	 */
    	@Column(name = "valid_flag" )
-	private Long validFlag;
+	private Integer validFlag;
 
 }
