+++ GSQ-forum-dao/src/main/java/com/sixtyrobbers/GSQ/fourm/dao/entity/fourm/dbdo/RoleDO.java
@@ -49,6 +49,6 @@ public class RoleDO  implements Serializable {
 	 * 有效标志位：0：存在；1：删除
 	 */
    	@Column(name = "valid_flag" )
-	private Long validFlag;
+	private Integer validFlag;
 
 }
